moves all creator component files into a single directory
