+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from './components/header'
 import Body from './components/body'
-import Creator from './components/creator'
+import Creator from './components/creator-module/creator'
 // import "./App.scss";
 
 function App() {
