+++ frontend/src/components/creator-module/creator-preview.scss
